refactor(taobao_guanghe): picker.py DOM 操作搬到 _link_ops
把 picker.py 中的 DOM 操作(scrape_products/scrape_shops/scrape_filters/
switch_radio/click_add_card/wait_panel_ready/switch_tab/click_filter/
search/load_more)搬到 _link_ops.py,picker.py 改为调 helper。纯重构,行为不变。
